+++ src/main/java/com/ssu/travel/domain/branch/service/BranchService.java
@@ -1,0 +1,4 @@
+package com.ssu.travel.domain.branch.service;
+
+public class BranchService {
+}
